+++ src/projectManager.tsx
@@ -1,0 +1,10 @@
+import * as React from "react";
+import { ProjectManager } from "./components/ProjectManager";
+
+/**
+ * Command entry point for Project Manager
+ * This file allows direct access to the project manager
+ */
+export default function Command() {
+  return <ProjectManager />;
+}