+++ src/utils/common.ts
@@ -4,6 +4,6 @@
 
 export interface Preferences {
   sonarqubePodmanDir: string;
-  useCustomSonarQubeApp: boolean;
   sonarqubeAppPath: string;
+  sonarqubePort?: string;
 }